pom.xml: Specify maven versions in properties
maven-project is pinned at 2.0.9, separate from maven.version as the
tooling module updates to maven 3.x which does not have a
maven-project.

Even though maven-project doesn't appear to be used and should
probably be removed, not pinning the version in dependencyManagement
will cause mvn to fail with an unresolved artifact.


Former-commit-id: 0033b11a1610a8c55356ff56f80c239515089023