KARAF-5314 Added a features cache that increases the performance of the filtering of the required features from the available features.
The performance of the filtering logic decresed when the java streams API was used in 4.1 compared to for loops in 4.0. I reverted the filtering logic to use for loops and also introduced a features cache. This increased the performance by a huge margin, almost 20X faster when there are ~900 available features and ~300 required features from a highly complex and huge feature dependency tree.


Former-commit-id: 3cbd810ea37e837950f1f8427f793c878423ba7a