KARAF-5604: Speed up features-generate-descriptor
OpenDaylight uses features-generate-descriptor to process a rather
large set (~196) of features being used as dependencies of a generated
feature. These features also contain a large number of bundles.

This patch introduces an explicit SimplLRUCache based on LinkedHashMap,
and places it into both GenerateDescriptorMojo and Dependency31Helper.
The size of the two cache instances can be controlled via plugin configuration
and default to 256 and 1024 entries respectively.

Signed-off-by: Robert Varga <nite@hq.sk>

Former-commit-id: 6d10058da48fe9846fb673aabd39a98dcadb2506